% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dparser.R
\name{.dparsePtr}
\alias{.dparsePtr}
\title{Return the dparser function pointers}
\usage{
.dparsePtr()
}
\value{
dparser function pointers
}
\description{
Return the dparser function pointers
}
\examples{
.dparsePtr()
}
\author{
Matthew L. Fidler
}
