% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dparser.R
\name{dparse_gram}
\alias{dparse_gram}
\title{R Parser for dparser grammar}
\usage{
dparse_gram(file, fn, envir = parent.frame(), start_state = 0,
  save_parse_tree = TRUE, partial_parses = FALSE, compare_stacks = TRUE,
  commit_actions_interval = 100, fixup = TRUE, fixup_ebnf = FALSE,
  nogreedy = FALSE, noheight = FALSE, use_file_name = TRUE,
  parse_size = 1024, verbose_level = 0)
}
\arguments{
\item{file}{File to be parsed}

\item{fn}{Function to be called at each production element.  This function should have the form fn(name,value,pos,depth).
\describe{

\item{\code{name}}{ represents the element name, usually a
production name.}

\item{\code{value}}{ reperesnts the value of the element at the
current position.}

\item{\code{pos}}{ is an integer reperesnting the token position.
When pos=-1, this is the whole token before looking at children.
Otherwise, the first part of the production terminal is numbered
from 0 to the end of the terminal. }

\item{\code{depth}}{ is an integer reperesnting the depth of parsing}

}}

\item{envir}{Environment where the function \code{fn} is called}

\item{start_state}{Start State (default 0)}

\item{save_parse_tree}{Save Parse Tree (default TRUE)}

\item{partial_parses}{Partial Parses (default FALSE)}

\item{compare_stacks}{Compare Stacks (default TRUE)}

\item{commit_actions_interval}{Commit Interval (default 100)}

\item{fixup}{Fixup Internal Productions (default FALSE)}

\item{fixup_ebnf}{Fixup EBNF Productions (default FALSE)}

\item{nogreedy}{No Greediness for Disambiguation (default FALSE)}

\item{noheight}{No Height for Disambiguation (default FALSE)}

\item{use_file_name}{Use File Name for syntax errors (default
TRUE)}

\item{parse_size}{Parser size (default 1024)}

\item{verbose_level}{the level of verbosity when creating parser
(default 0)}
}
\value{
Nothing is returned
}
\description{
This is a parser that takes a dparser grammer and walks the
grammar function.
}
\author{
Matthew L. Fidler
}

