% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dparser.R
\docType{class}
\name{dparserFunction-class}
\alias{dparserFunction-class}
\title{A S4 class for dparser functions}
\description{
The data here is the function to be called.
}
\section{Slots}{

\describe{
\item{\code{env}}{contains the environment where the grammar name and dll
files are stored.}

\item{\code{.Data}}{is the function data}
}}

\keyword{internal}
