% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dparser.R
\name{dpDefaultSkip}
\alias{dpDefaultSkip}
\title{Default skip function for darsing grammar}
\usage{
dpDefaultSkip(name, value, pos, depth)
}
\arguments{
\item{name}{Production Name}

\item{value}{Production Value}

\item{pos}{terminal position.  Negative values are parsed before
children are parsed.  A value of -1 means there are no
children nodes.  A value of -2 means there are children nodes.
Otherwise, the terminal position is numbered starting at 0.}

\item{depth}{Parsing Depth}
}
\value{
FALSE.  If a comparable function returns TRUE, the
    children are not parsed.
}
\description{
This function is to determine if children are parsed or skipped.
By default, all children are parsed.
}
\author{
Matthew L. Fidler
}
