% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dparser.R
\name{dpGetFile}
\alias{dpGetFile}
\title{Get file for dparser arguments}
\usage{
dpGetFile(file, fileext = "")
}
\arguments{
\item{file}{- File to handle.  This file can be:
\itemize{

\item an actual file.

\item a character string.  In this case, a temporary file is
created with the character string.

\item A bracket \code{\{\}} enclosed R expression.  In this case
the contents of the expressions is put into a temporary file.

\item A function.  In this case the contents of the function body are put into the temporary file.

}}

\item{fileext}{file extension of temporary file to create}
}
\value{
a list of two elements:
\describe{

\item{\code{file}}{The file name of etiher the tempoary file or the real file}

\item{code{use_file_name}}{If the file name was used (\code{TRUE}), or a temporary file was created(\code{FALSE})}

}
}
\description{
Get file for dparser arguments
}
\author{
Matthew L. Fidler
}
\keyword{internal}

