% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dpVersion}
\alias{dpVersion}
\title{Version and repository for this dparser package.}
\usage{
dpVersion()
}
\value{
A character vector with the version and repository.
}
\description{
Version and repository for this dparser package.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
