% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparse.R
\name{dpRparse}
\alias{dpRparse}
\title{C code for R parser}
\usage{
dpRparse()
}
\value{
A string for generating C tree walker
}
\description{
C code for R parser
}
\author{
Matthew L. Fidler
}
\keyword{internal}
